hellofinitestates: refactored `TransitionalManager` to more compatible OOP principles
