+++ HelloFiniteStateAutomaton/app/src/main/java/com/scrappers/fsa/core/TransitionalManager.java
@@ -7,38 +7,22 @@ import java.lang.Thread;
 import com.scrappers.fsa.core.state.AutoState;
 import com.scrappers.fsa.core.state.TransitionListener;
 
-public final class TransitionalManager {
+public class TransitionalManager {
     
     public static final Logger LOGGER = Logger.getLogger(TransitionalManager.class.getName());
     public final List<AutoState<?, ?>> autostates = new ArrayList<>();
-    
-    private static TransitionalManager INSTANCE;
-    
-    private TransitionalManager() {
-    }
-    
-    public static TransitionalManager initialize() {
-        synchronized (TransitionalManager.class) {
-            /* sanity check the input */
-            if (INSTANCE != null) {
-                return INSTANCE;
-            }
-            INSTANCE = new TransitionalManager();
-        }
-        return INSTANCE;
-    }
      
-    public void assignNextState(AutoState<?, ?> autostate) { 
+    public <I, O> void assignNextState(AutoState<I, O> autostate) { 
         autostates.add(autostate);
     }
     
-    public <T> void transit(final long time, final T input, final TransitionListener transitionListener) throws InterruptedException {
+    public <I> void transit(final long time, final I input, final TransitionListener transitionListener) throws InterruptedException {
         Thread.sleep(time);
-        transitNow(input, transitionListener);
+        transit(input, transitionListener);
     }
     
-    public <I> void transitNow(final I input, final TransitionListener transitionListener) {
-        final AutoState<I, ?> autoState = (AutoState<I, ?>) autostates.iterator().next();
+    public <I, O> void transit(final I input, final TransitionListener transitionListener) {
+        final AutoState<I, O> autoState = (AutoState<I, O>) autostates.iterator().next();
         /* sanity check the input */
         assert autostates.contains(input);
         autoState.invoke(input);
