+++ HelloFiniteStateAutomaton/app/src/main/java/com/scrappers/fsa/core/package-info.java
@@ -1,1 +1,1 @@
-package com.scrappers.fsa;
+package com.scrappers.fsa.core;
